Several Bug Fixes (#23)
* Fix rigidbody falling detection not being properly updated

falling detection was only being called for the rigidbody which the handleCollision was called in. Moving this detection to respondToCollision ensures that both rigidbodies involved in a collision respond accurately

* use an epsolon

* dereference pointer

* dont return a positive collision if mtv is a 0 vector

* minimal performance improvement

* fix falling detection

only move into a non-falling state if a collision has occurred and the velocity is down or constant and the collision is along the y axis, or if the previous velocity was < 0 and an applied force has changes it to be > 0.